+++ examples/bond/test/bond-token-flow.js
@@ -1,4 +1,4 @@
-const { getProgram, convert } = require('@algo-builder/algob');
+const { convert } = require('@algo-builder/algob');
 const {
   Runtime, AccountStore
 } = require('@algo-builder/runtime');
@@ -107,8 +107,7 @@ describe('Bond token Tests', function () {
       TMPL_OWNER: bondTokenCreator.address,
       TMPL_APP_MANAGER: appManager.address
     };
-    const issuerLsigProg = getProgram('issuer-lsig.py', scInitParam);
-    lsig = runtime.createLsigAccount(issuerLsigProg, []);
+    lsig = runtime.loadLogic('issuer-lsig.py', scInitParam);
     issuerLsigAddress = lsig.address();
 
     // sync escrow account
@@ -246,8 +245,7 @@ describe('Bond token Tests', function () {
       TMPL_APP_MANAGER: appManager.address,
       TMPL_BOND: bond2
     };
-    const buyLsigProgram = getProgram('buyback-lsig.py', scParam);
-    const buybackLsig = runtime.createLsigAccount(buyLsigProgram, []);
+    const buybackLsig = runtime.loadLogic('buyback-lsig.py', scParam);
 
     // fund dex with some minimum balance first
     runtime.fundLsig(master.account, buybackLsig.address(), minBalance + 10000);