+++ examples/bond/test/failing-tests.js
@@ -1,4 +1,4 @@
-const { getProgram, convert } = require('@algo-builder/algob');
+const { convert } = require('@algo-builder/algob');
 const {
   Runtime, AccountStore
 } = require('@algo-builder/runtime');
@@ -95,8 +95,7 @@ describe('Bond token failing tests', function () {
       TMPL_OWNER: bondTokenCreator.address,
       TMPL_APP_MANAGER: appManager.address
     };
-    const issuerLsigProg = getProgram('issuer-lsig.py', scInitParam);
-    lsig = runtime.createLsigAccount(issuerLsigProg, []);
+    lsig = runtime.loadLogic('issuer-lsig.py', scInitParam);
     issuerLsigAddress = lsig.address();
 
     // sync escrow account