+++ examples/crowdfunding/test/crowdfundingTest.js
@@ -1,4 +1,5 @@
-const { getProgram, convert } = require('@algo-builder/algob');
+const { getProgram } = require('@algo-builder/runtime');
+const { convert } = require('@algo-builder/algob');
 const {
   Runtime, AccountStore
 } = require('@algo-builder/runtime');
@@ -97,8 +98,7 @@ describe('Crowdfunding Tests', function () {
     const creatorPk = convert.addressToPk(creator.address);
 
     // setup escrow account
-    const escrowProg = getProgram('crowdFundEscrow.py', { APP_ID: applicationId });
-    const lsig = runtime.createLsigAccount(escrowProg, []);
+    const lsig = runtime.loadLogic('crowdFundEscrow.py', { APP_ID: applicationId });
     const escrowAddress = lsig.address();
 
     // sync escrow account
@@ -300,8 +300,7 @@ describe('Crowdfunding Tests', function () {
       { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram).appID;
 
     // setup escrow account
-    const escrowProg = getProgram('crowdFundEscrow.py', { APP_ID: applicationId });
-    const lsig = runtime.createLsigAccount(escrowProg, []);
+    const lsig = runtime.loadLogic('crowdFundEscrow.py', { APP_ID: applicationId });
     const escrowAddress = lsig.address();
 
     // sync escrow account