+++ examples/permissioned-token/test/common.js
@@ -1,7 +1,4 @@
-const {
-  getProgram
-} = require('@algo-builder/algob');
-const { Runtime } = require('@algo-builder/runtime');
+const { Runtime, getProgram } = require('@algo-builder/runtime');
 const { types } = require('@algo-builder/web');
 
 const minBalance = 20e6; // 20 ALGOs
@@ -75,11 +72,10 @@ class Context {
 
   // Deploy Clawback Lsig and Modify Asset
   deployClawback (sender, clawbackProgram) {
-    const clawbackTeal = getProgram(clawbackProgram, {
+    this.lsig = this.runtime.loadLogic(clawbackProgram, {
       TOKEN_ID: this.assetIndex,
       CONTROLLER_APP_ID: this.controllerappID
     });
-    this.lsig = this.runtime.createLsigAccount(clawbackTeal, []);
 
     fund(this.runtime, this.master, this.lsig.address());
     const asaDef = this.runtime.getAssetDef(this.assetIndex);