+++ packages/algob/src/lib/compile.ts
@@ -1,7 +1,6 @@
-import { getPathFromDirRecursive } from "@algo-builder/runtime";
-import { BuilderError, ERRORS, parseAlgorandError, types } from "@algo-builder/web";
+import { getPathFromDirRecursive, PyCompileOp } from "@algo-builder/runtime";
+import { parseAlgorandError, types } from "@algo-builder/web";
 import type { Algodv2, modelsv2 } from "algosdk";
-import { spawnSync, SpawnSyncReturns } from "child_process";
 import * as fs from 'fs';
 import * as murmurhash from 'murmurhash';
 import * as path from 'path';
@@ -9,7 +8,7 @@ import YAML from "yaml";
 
 import { assertDir, ASSETS_DIR, CACHE_DIR } from "../internal/core/project-structure";
 import { timestampNow } from "../lib/time";
-import type { ASCCache, PyASCCache, ReplaceParams, SCParams } from "../types";
+import type { ASCCache, PyASCCache, SCParams } from "../types";
 
 export const tealExt = ".teal";
 export const pyExt = ".py";
@@ -22,7 +21,7 @@ export class CompileOp {
 
   constructor (algocl: Algodv2) {
     this.algocl = algocl;
-    this.pyCompile = new PyCompileOp(this);
+    this.pyCompile = new PyCompileOp();
   }
 
   /** Gets the TEAL compiled result from artifacts cache and compiles the code if necessary.
@@ -33,22 +32,28 @@ export class CompileOp {
    * @param force: if true it will force recompilation even if the cache is up to date.
    * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
    */
-  async ensureCompiled (filename: string, force?: boolean, scTmplParams?: SCParams): Promise<ASCCache> {
+  async ensureCompiled (filename: string, force?: boolean, scTmplParams?: SCParams):
+  Promise<ASCCache | PyASCCache> {
     const filePath = getPathFromDirRecursive(ASSETS_DIR, filename) as string;
 
     if (force === undefined) {
       force = false;
     }
 
+    if (!filename.endsWith(tealExt) && !filename.endsWith(lsigExt) && !filename.endsWith(pyExt)) {
+      throw new Error(`filename "${filename}" must end with "${tealExt}" or "${lsigExt}" or "${pyExt}"`); // TODO: convert to buildererror
+    }
+
+    let teal: string;
+    let thash: number;
     if (filename.endsWith(pyExt)) {
-      return await this.pyCompile.ensureCompiled(filename, force, scTmplParams);
+      const content = this.pyCompile.ensurePyTEALCompiled(filename, scTmplParams);
+      [teal, thash] = [content, murmurhash.v3(content)];
+    } else {
+      [teal, thash] = this.readTealAndHash(filePath);
     }
 
-    if (!filename.endsWith(tealExt) && !filename.endsWith(lsigExt)) {
-      throw new Error(`filename "${filename}" must end with "${tealExt}" or "${lsigExt}"`); // TODO: convert to buildererror
-    }
-
-    const [teal, thash] = this.readTealAndHash(filePath);
+    // const [teal, thash] = this.readTealAndHash(filePath);
     let a = await this.readArtifact(filename);
     if (!force && a !== undefined && a.srcHash === thash) {
       // '\x1b[33m%s\x1b[0m' for yellow color warning
@@ -90,7 +95,7 @@ export class CompileOp {
     return this.algocl.compile(code).do();
   }
 
-  async compile (filename: string, tealCode: string, tealHash: number): Promise<ASCCache> {
+  async compile (filename: string, tealCode: string, tealHash: number): Promise<ASCCache | PyASCCache> {
     try {
       const co = await this.callCompiler(tealCode);
       return {
@@ -99,6 +104,7 @@ export class CompileOp {
         compiled: co.result,
         compiledHash: co.hash,
         srcHash: tealHash,
+        tealCode: filename.endsWith(pyExt) ? tealCode : undefined,
         // compiled base64 converted into bytes
         base64ToBytes: new Uint8Array(Buffer.from(co.result, "base64"))
       };
@@ -110,146 +116,5 @@ export class CompileOp {
 
   writeFile (filename: string, content: string): void {
     fs.writeFileSync(filename, content);
-  }
-}
-
-export class PyCompileOp {
-  compileOp: CompileOp;
-
-  constructor (compileOp: CompileOp) {
-    this.compileOp = compileOp;
-  }
-
-  /**
-   * Parses scTmplParams and returns ReplaceParams and stringify object
-   * @param scTmplParams smart contract template parameters
-   */
-  parseScTmplParam (scTmplParams?: SCParams): [ReplaceParams, string | undefined] {
-    let param: string | undefined;
-    const replaceParams: ReplaceParams = {};
-    if (scTmplParams === undefined) {
-      param = undefined;
-    } else {
-      const tmp: SCParams = {};
-      for (const key in scTmplParams) {
-        if (key.startsWith("TMPL_") || key.startsWith("tmpl_")) {
-          replaceParams[key] = scTmplParams[key].toString();
-        } else {
-          tmp[key] = scTmplParams[key];
-        }
-      }
-      console.log("PyTEAL template parameters:", tmp);
-      param = YAML.stringify(tmp);
-    }
-    console.log("TEAL replacement parameters:", replaceParams);
-    return [replaceParams, param];
-  }
-
-  /**
-   * Description : returns compiled teal code from pyTeal file
-   * @param filename : name of the PyTeal code in `/assets` directory.
-   *                   Examples : [ gold.py, asa.py]
-   *                   MUST have .py extension
-   * @param force    : if true it will force recompilation even if the cache is up to date.
-   * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
-   */
-  async ensureCompiled (filename: string, force?: boolean, scTmplParams?: SCParams): Promise<PyASCCache> {
-    if (!filename.endsWith(pyExt)) {
-      throw new Error(`filename "${filename}" must end with "${pyExt}"`);
-    }
-
-    const [replaceParams, param] = this.parseScTmplParam(scTmplParams);
-    let content = this.compilePyTeal(filename, param);
-    if (YAML.stringify({}) !== YAML.stringify(replaceParams)) {
-      content = this.replaceTempValues(content, replaceParams);
-    }
-    const [teal, thash] = [content, murmurhash.v3(content)];
-
-    const a = await this.readArtifact(filename);
-    if (!force && a !== undefined && a.srcHash === thash) {
-      // '\x1b[33m%s\x1b[0m' for yellow color warning
-      console.warn('\x1b[33m%s\x1b[0m', `smart-contract source "${filename}" didn't change, skipping.`);
-      return a;
-    }
-    console.log("compiling", filename);
-    const compiledTeal = await this.compileOp.compile(filename, teal, thash);
-    const pyCompiled: PyASCCache = {
-      filename: "",
-      timestamp: 0,
-      compiled: "",
-      compiledHash: "",
-      srcHash: 0,
-      tealCode: content,
-      base64ToBytes: new Uint8Array(1)
-    };
-    Object.assign(pyCompiled, compiledTeal);
-
-    const cacheFilename = path.join(CACHE_DIR, filename + ".yaml");
-    this.compileOp.writeFile(cacheFilename, YAML.stringify(pyCompiled));
-    return pyCompiled;
-  }
-
-  /**
-   * Replaces keys with the values in program using replaceParams
-   * @param program Teal program in string
-   * @param replaceParams params that needs to be replaced in program
-   */
-  replaceTempValues (program: string, replaceParams: ReplaceParams): string {
-    for (const param in replaceParams) {
-      program = program.split(param).join(replaceParams[param]);
-    }
-    return program;
-  }
-
-  async readArtifact (filename: string): Promise<PyASCCache | undefined> {
-    await assertDir(CACHE_DIR);
-    try {
-      const p = path.join(CACHE_DIR, filename + ".yaml");
-      return YAML.parse(await fs.promises.readFile(p, 'utf8')) as PyASCCache;
-    } catch (e) {
-      if (types.isFileError(e) && e?.errno === -2) { return undefined; } // handling a not existing file
-      throw e;
-    }
-  }
-
-  /**
-   * Description: Runs a subprocess to execute python script
-   * @param filename : python filename in assets folder
-   * @param scInitParam : Smart contract initialization parameters.
-   */
-  private runPythonScript (filename: string, scInitParam?: string): SpawnSyncReturns<string> {
-    const filePath = getPathFromDirRecursive(ASSETS_DIR, filename) as string;
-    // used spawnSync instead of spawn, as it is synchronous
-    if (scInitParam === undefined) {
-      return spawnSync(
-        'python3', [
-          filePath
-        ], { encoding: 'utf8' }
-      );
-    }
-
-    return spawnSync('python3', [
-      filePath,
-      scInitParam
-    ], { encoding: 'utf8' }
-    );
-  }
-
-  /**
-   * Description: returns TEAL code using pyTeal compiler
-   * @param filename : python filename in assets folder
-   * @param scInitParam : Smart contract initialization parameters.
-   */
-  compilePyTeal (filename: string, scInitParam?: string): string {
-    const subprocess: SpawnSyncReturns<string> = this.runPythonScript(filename, scInitParam);
-
-    if (subprocess.stderr) {
-      throw new BuilderError(
-        ERRORS.PyTEAL.PYTEAL_FILE_ERROR, {
-          filename: filename,
-          reason: subprocess.stderr
-        });
-    }
-    return subprocess.stdout;
   }
 }