+++ packages/runtime/test/integration/app-update.ts
@@ -1,10 +1,10 @@
 import { types } from "@algo-builder/web";
 import { assert } from "chai";
 
+import { getProgram } from "../../src";
 import { RUNTIME_ERRORS } from "../../src/errors/errors-list";
 import { AccountStore, Runtime } from "../../src/index";
 import { ALGORAND_ACCOUNT_MIN_BALANCE } from "../../src/lib/constants";
-import { getProgram } from "../helpers/files";
 import { useFixture } from "../helpers/integration";
 import { expectRuntimeError } from "../helpers/runtime-errors";
 
@@ -22,7 +22,7 @@ describe("App Update Test", function () {
 
   this.beforeEach(async function () {
     runtime = new Runtime([john, alice]); // setup test
-    approvalProgram = getProgram('approval_program.teal');
+    approvalProgram = getProgram('approval_program.py');
     clearProgram = getProgram('clear_program.teal');
     const flags = {
       sender: john.account,