+++ packages/runtime/test/integration/close-clear-ssc.ts
@@ -2,10 +2,10 @@ import { parsing, types } from "@algo-builder/web";
 import { assert } from "chai";
 
 import { SSC_VALUE_BYTES } from "../../build/lib/constants";
+import { getProgram } from "../../src";
 import { RUNTIME_ERRORS } from "../../src/errors/errors-list";
 import { AccountStore, Runtime } from "../../src/index";
 import { ALGORAND_ACCOUNT_MIN_BALANCE, APPLICATION_BASE_FEE, SSC_VALUE_UINT } from "../../src/lib/constants";
-import { getProgram } from "../helpers/files";
 import { useFixture } from "../helpers/integration";
 import { expectRuntimeError } from "../helpers/runtime-errors";